+++ packages/charts/src/chart_types/heatmap/layout/viewmodel/viewmodel.ts
@@ -92,6 +92,7 @@ export function shapeViewModel<D extends BaseDatum = Datum>(
   const boxedYValues = yValues.map<Box & { value: NonNullable<PrimitiveValue> }>((value) => ({
     text: config.yAxisLabel.formatter(value),
     value,
+    isValue: false,
     ...config.yAxisLabel,
   }));
 
@@ -399,6 +400,7 @@ function getXTicks(
     return {
       text: formatter(value),
       value,
+      isValue: false,
       ...config.xAxisLabel,
       x: chartDimensions.left + (scaleCallback(value) ?? 0),
       y: gridHeight + config.xAxisLabel.fontSize / 2 + config.xAxisLabel.padding,