+++ packages/charts/src/chart_types/xy_chart/renderer/canvas/xy_chart.tsx
@@ -50,9 +50,11 @@ import { LinesGrid } from '../../utils/grid_lines';
 import { IndexedGeometryMap } from '../../utils/indexed_geometry_map';
 import { AxisSpec, AnnotationSpec } from '../../utils/specs';
 import { renderXYChartCanvas2d } from './renderers';
+import { hasMostlyRTL } from './utils/has_mostly_rtl';
 
 /** @internal */
 export interface ReactiveChartStateProps {
+  isRTL: boolean;
   initialized: boolean;
   debug: boolean;
   isChartEmpty: boolean;
@@ -151,6 +153,7 @@ class XYChartComponent extends React.Component<XYChartProps> {
       chartContainerDimensions: { width, height },
       a11ySettings,
       debug,
+      isRTL,
     } = this.props;
 
     if (!initialized || isChartEmpty) {
@@ -162,6 +165,7 @@ class XYChartComponent extends React.Component<XYChartProps> {
       <>
         <figure aria-labelledby={a11ySettings.labelId} aria-describedby={a11ySettings.descriptionId}>
           <canvas
+            dir={isRTL ? 'rtl' : 'ltr'}
             ref={forwardCanvasRef}
             className="echCanvasRenderer"
             width={width * this.devicePixelRatio}
@@ -190,6 +194,7 @@ const mapDispatchToProps = (dispatch: Dispatch): ReactiveChartDispatchProps =>
   );
 
 const DEFAULT_PROPS: ReactiveChartStateProps = {
+  isRTL: false,
   initialized: false,
   debug: false,
   isChartEmpty: true,
@@ -238,8 +243,10 @@ const mapStateToProps = (state: GlobalChartState): ReactiveChartStateProps => {
 
   const { geometries, geometriesIndex } = computeSeriesGeometriesSelector(state);
   const { debug } = getSettingsSpecSelector(state);
+  const perPanelAxisGeoms = computePerPanelAxesGeomsSelector(state);
 
   return {
+    isRTL: hasMostlyRTL(perPanelAxisGeoms),
     initialized: true,
     isChartEmpty: isChartEmptySelector(state),
     debug,
@@ -252,7 +259,7 @@ const mapStateToProps = (state: GlobalChartState): ReactiveChartStateProps => {
     renderingArea: computeChartDimensionsSelector(state).chartDimensions,
     chartTransform: computeChartTransformSelector(state),
     axesSpecs: getAxisSpecsSelector(state),
-    perPanelAxisGeoms: computePerPanelAxesGeomsSelector(state),
+    perPanelAxisGeoms,
     perPanelGridLines: getGridLinesSelector(state),
     axesStyles: getAxesStylesSelector(state),
     annotationDimensions: computeAnnotationDimensionsSelector(state),