+++ packages/charts/src/chart_types/xy_chart/rendering/area.ts
@@ -36,13 +36,13 @@ export function renderArea(
   panel: Dimensions,
   color: Color,
   curve: CurveType,
-  hasY0Accessors: boolean,
+  isBandedSpec: boolean,
   xScaleOffset: number,
-  seriesStyle: AreaSeriesStyle,
+  style: AreaSeriesStyle,
   markSizeOptions: MarkSizeOptions,
-  isStacked = false,
+  isStacked: boolean,
+  hasFit: boolean,
   pointStyleAccessor?: PointStyleAccessor,
-  hasFit?: boolean,
 ): {
   areaGeometry: AreaGeometry;
   indexedGeometryMap: IndexedGeometryMap;
@@ -57,14 +57,15 @@ export function renderArea(
     .y1(y1Fn)
     .y0(y0Fn)
     .defined((datum) => {
-      return definedFn(datum, y1DatumAccessor) && (hasY0Accessors ? definedFn(datum, y0DatumAccessor) : true);
+      return definedFn(datum, y1DatumAccessor) && (isBandedSpec ? definedFn(datum, y0DatumAccessor) : true);
     })
     .curve(getCurveFactory(curve));
 
+  // TODO we can probably avoid this function call if no fit function is applied.
   const clippedRanges = getClippedRanges(dataSeries.data, xScale, xScaleOffset);
 
   const lines: string[] = [];
-  const y0Line = hasY0Accessors && pathGenerator.lineY0()(dataSeries.data);
+  const y0Line = isBandedSpec && pathGenerator.lineY0()(dataSeries.data);
   const y1Line = pathGenerator.lineY1()(dataSeries.data);
   if (y1Line) lines.push(y1Line);
   if (y0Line) lines.push(y0Line);
@@ -76,11 +77,11 @@ export function renderArea(
     yScale,
     panel,
     color,
-    seriesStyle.point,
-    hasY0Accessors,
+    style.point,
+    isBandedSpec,
     markSizeOptions,
-    pointStyleAccessor,
     false,
+    pointStyleAccessor,
   );
 
   const areaGeometry: AreaGeometry = {
@@ -101,12 +102,10 @@ export function renderArea(
       smHorizontalAccessorValue: dataSeries.smHorizontalAccessorValue,
       smVerticalAccessorValue: dataSeries.smVerticalAccessorValue,
     },
-    seriesAreaStyle: seriesStyle.area,
-    seriesAreaLineStyle: seriesStyle.line,
-    seriesPointStyle: seriesStyle.point,
+    style,
     isStacked,
     clippedRanges,
-    hideClippedRanges: !hasFit,
+    shouldClip: hasFit,
   };
   return {
     areaGeometry,