+++ packages/charts/src/chart_types/xy_chart/utils/series.test.ts
@@ -27,10 +27,10 @@ import {
   getSeriesColors,
   getDataSeriesFromSpecs,
   XYChartSeriesIdentifier,
-  extractYAndMarkFromDatum,
   getSeriesName,
   DataSeries,
   splitSeriesDataByAccessors,
+  extractYAndMarkFromDatum,
 } from './series';
 import { BasicSeriesSpec, LineSeriesSpec, SeriesType, AreaSeriesSpec } from './specs';
 import { formatStackedDataSeriesValues } from './stacked_series_utils';
@@ -236,7 +236,7 @@ describe('Series', () => {
       }),
     ];
     const xValues = new Set([1, 2, 3, 4]);
-    const stackedValues = formatStackedDataSeriesValues(dataSeries, xValues);
+    const stackedValues = formatStackedDataSeriesValues(dataSeries, xValues, 'bar');
     expect(stackedValues.map(matchOnlyDataSeriesLegacySnapshot)).toMatchSnapshot();
   });
   test('Can stack unsorted dataseries', () => {
@@ -287,7 +287,7 @@ describe('Series', () => {
       }),
     ];
     const xValues = new Set(new Array(maxArrayItems).fill(0).map((d, i) => i));
-    const stackedValues = formatStackedDataSeriesValues(dataSeries, xValues);
+    const stackedValues = formatStackedDataSeriesValues(dataSeries, xValues, 'bar');
     expect(stackedValues.map(matchOnlyDataSeriesLegacySnapshot)).toMatchSnapshot();
   });
   test('Can stack simple dataseries with scale to extent', () => {
@@ -599,26 +599,26 @@ describe('Series', () => {
   });
 
   test('clean datum shall parse string as number for y values', () => {
-    let datum = extractYAndMarkFromDatum([0, 1, 2], 1, [], 2);
+    let datum = extractYAndMarkFromDatum([0, 1, 2], 1, [], false, 2);
     expect(datum).toBeDefined();
     expect(datum?.y1).toBe(1);
-    expect(datum?.y0).toBe(2);
-    datum = extractYAndMarkFromDatum([0, '1', 2], 1, [], 2);
+    expect(datum?.y0).toBe(null);
+    datum = extractYAndMarkFromDatum([0, '1', 2], 1, [], false, 2);
     expect(datum).toBeDefined();
     expect(datum?.y1).toBe(1);
-    expect(datum?.y0).toBe(2);
+    expect(datum?.y0).toBe(null);
 
-    datum = extractYAndMarkFromDatum([0, '1', '2'], 1, [], 2);
+    datum = extractYAndMarkFromDatum([0, '1', '2'], 1, [], false, 2);
     expect(datum).toBeDefined();
     expect(datum?.y1).toBe(1);
-    expect(datum?.y0).toBe(2);
+    expect(datum?.y0).toBe(null);
 
-    datum = extractYAndMarkFromDatum([0, 1, '2'], 1, [], 2);
+    datum = extractYAndMarkFromDatum([0, 1, '2'], 1, [], false, 2);
     expect(datum).toBeDefined();
     expect(datum?.y1).toBe(1);
-    expect(datum?.y0).toBe(2);
+    expect(datum?.y0).toBe(null);
 
-    datum = extractYAndMarkFromDatum([0, 'invalid', 'invalid'], 1, [], 2);
+    datum = extractYAndMarkFromDatum([0, 'invalid', 'invalid'], 1, [], false, 2);
     expect(datum).toBeDefined();
     expect(datum?.y1).toBe(null);
     expect(datum?.y0).toBe(null);