+++ packages/charts/src/chart_types/xy_chart/utils/series.ts
@@ -134,6 +134,12 @@ export function splitSeriesDataByAccessors(
   const xValues: Array<string | number> = [];
   const nonNumericValues: any[] = [];
 
+  if (isStacked && Boolean(y0Accessors?.length)) {
+    Logger.warn(
+      `y0Accessors are not allowed with stackAccessors. y0Accessors will be ignored but available under initialY0.`,
+    );
+  }
+
   for (let i = 0; i < data.length; i++) {
     const datum = data[i];
     const splitAccessors = getSplitAccessors(datum, splitSeriesAccessors);
@@ -164,6 +170,7 @@ export function splitSeriesDataByAccessors(
         datum,
         accessor,
         nonNumericValues,
+        isBandedSpec(spec),
         y0Accessors && y0Accessors[index],
         markSizeAccessor,
       );
@@ -273,6 +280,7 @@ export function extractYAndMarkFromDatum<D extends BaseDatum>(
   datum: D,
   yAccessor: Accessor<D> | AccessorFn<D>,
   nonNumericValues: any[],
+  bandedSpec: boolean,
   y0Accessor?: Accessor<D> | AccessorFn<D>,
   markSizeAccessor?: Accessor<D> | AccessorFn<D>,
 ): Pick<DataSeriesDatum, 'y0' | 'y1' | 'mark' | 'datum' | 'initialY0' | 'initialY1'> {
@@ -281,7 +289,7 @@ export function extractYAndMarkFromDatum<D extends BaseDatum>(
   const y1Value = getAccessorValue(datum, yAccessor);
   const y1 = finiteOrNull(y1Value, nonNumericValues);
   const y0 = y0Accessor ? finiteOrNull(getAccessorValue(datum, y0Accessor), nonNumericValues) : null;
-  return { y1, datum, y0, mark, initialY0: y0, initialY1: y1 };
+  return { y1, datum, y0: bandedSpec ? y0 : null, mark, initialY0: y0, initialY1: y1 };
 }
 
 function finiteOrNull(value: unknown, nonNumericValues: unknown[]): number | null {
@@ -327,8 +335,8 @@ export function getFormattedDataSeries(
   );
 
   const fittedAndStackedDataSeries = stackedGroups.reduce<DataSeries[]>((acc, dataSeries) => {
-    const [{ stackMode }] = dataSeries;
-    const formatted = formatStackedDataSeriesValues(dataSeries, xValues, stackMode);
+    const [{ stackMode, seriesType }] = dataSeries;
+    const formatted = formatStackedDataSeriesValues(dataSeries, xValues, seriesType, stackMode);
     return [...acc, ...formatted];
   }, []);
   // get already fitted non stacked dataSeries
@@ -447,9 +455,12 @@ export function getDataSeriesFromSpecs(
   };
 }
 
-/** @internal */
-export function isDataSeriesBanded({ spec }: DataSeries) {
-  return spec.y0Accessors && spec.y0Accessors.length > 0;
+/**
+ * TODO: Add check for chart type other than area and bar.
+ * @internal
+ */
+export function isBandedSpec(spec: BasicSeriesSpec): boolean {
+  return Boolean(spec.y0Accessors && spec.y0Accessors.length > 0 && !isStackedSpec(spec, false));
 }
 
 function getSortedOrdinalXValues(